refactor(react): rename `/adapters` subpath to `/loaders`
