+++ .changeset/loaders-subpath-rename.md
@@ -1,0 +1,5 @@
+---
+'@bazza-ui/react': minor
+---
+
+Rename the `@bazza-ui/react/adapters` subpath to `@bazza-ui/react/loaders`. The optional `QueryDependentLoaderConfig` type is now `QueryLoaderConfig`; update imports accordingly.
