+++ packages/react/src/internal/popup-menu/deep-search/data-list.tsx
@@ -44,7 +44,7 @@ import type {
   GroupRenderContext,
   ItemDef,
   NodeDef,
-  QueryDependentLoaderConfig,
+  QueryLoaderConfig,
   RadioGroupDef,
   RowRenderContext,
   SubmenuDef,
@@ -223,7 +223,7 @@ interface QueryExecutionState {
   isBelowMinLength: boolean
 }
 
-function resolveInitialQueryBehavior(config: QueryDependentLoaderConfig):
+function resolveInitialQueryBehavior(config: QueryLoaderConfig):
   | {
       value: string
       loadWhen: 'needed' | 'parent-open'
@@ -247,7 +247,7 @@ function resolveInitialQueryBehavior(config: QueryDependentLoaderConfig):
 }
 
 function resolveQueryExecutionState(
-  config: QueryDependentLoaderConfig,
+  config: QueryLoaderConfig,
   query: string,
 ): QueryExecutionState {
   const minLength = config.minQueryLength ?? 1
