+++ packages/react/src/loaders/swr.test.ts
