+++ packages/react/src/loaders/vanilla.test.tsx
